Fix NPP_Export x64 crash on Notepad++ v8.3
And make it compilable on VS2017

+++ src/Notepad_specific/Scintilla.h
@@ -1063,8 +1063,8 @@ namespace Scintilla {
 #endif
 
 struct Sci_CharacterRange {
-	long cpMin;
-	long cpMax;
+	intptr_t cpMin;
+	intptr_t cpMax;
 };
 
 struct Sci_TextRange {
